% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iQualityR-package.R
\name{iQualityR_packages}
\alias{iQualityR_packages}
\title{List iQualityR member packages}
\usage{
iQualityR_packages()
}
\value{
A \code{data.frame} with three columns:
\describe{
\item{package}{Character. Name of the member package.}
\item{title}{Character. One-line description from the package DESCRIPTION.}
\item{version}{Character. Installed version of the member package.}
}
}
\description{
Returns a data frame describing the member packages that are
attached when \code{library(iQualityR)} is called. Useful for
introspection, troubleshooting, and reporting.
}
\examples{
iQualityR_packages()

}
